Create feature for add products to shopping cart
This features use a service which implement a BehaviorSubject, it has an
observable and the addCart method which send a notification to
subscribers.
Product component uses addCart method of the service to add a single
product to cart.
Header component subscribes to cart observable to get the notification
of a new product.